refactor(render): split node/window modules and rename helpers
Reorganize render modules to clearly separate node and window concerns,
and rename shared helpers for consistency.

- restructure render tree into `cursor.rs`, `frame.rs`, `layer_shell.rs`,
  `node.rs`, `window.rs`, and `utils.rs`
- move window-related rendering (active windows, popups, offscreen
  collection) into `window.rs`
- move node-related rendering (shaders, markers, hover previews, labels)
  into `node.rs`
- update imports across `mod.rs`, `frame.rs`, `app_icon.rs`, and other
  render code to match new boundaries
- remove legacy `node_render.rs` and update remaining `render_utils`
  references after rename to `utils.rs`
